lint fixes

+++ src/reviews/reviews.service.ts
@@ -1,14 +1,20 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { ReviewRepository } from './review.repository';
 import { CreateReviewDto, ReviewFilterDto } from './dto';
 import { Review } from './entities/reviews.entity';
+import { AuthService } from '../auth/services';
 
 @Injectable()
 export class ReviewsService {
   constructor(
     @InjectRepository(ReviewRepository)
     private reviewRepository: ReviewRepository,
+    private authService: AuthService,
   ) {}
 
   async getStats(userId: number): Promise<any> {
@@ -45,6 +51,8 @@ export class ReviewsService {
     if (!user.id) {
       return;
     }
+    const access = await this.authService.getUserById(req.user.id);
+    if (access.role === 'admin') {
     const review = await this.reviewRepository.findOne({ id });
     if (!review) {
       throw new NotFoundException(`review with ${id} not found`);
@@ -56,16 +64,30 @@ export class ReviewsService {
       message: 'Status Updated Successfully',
       review: review,
     };
-  }
-  async deleteReview(id: number): Promise<any> {
-    const result = await this.reviewRepository.delete({ id });
-
-    if (result.affected === 0) {
-      throw new NotFoundException(`Review with ${id} not found`);
+  } else{
+      throw new UnauthorizedException(
+          'UnAuthorized',
+          'You are not Authorized to access this endpoint ',
+      );
     }
-    return {
-      success: true,
-      message: 'Review deleted Successfully',
-    };
+  }
+  async deleteReview(id: number, req: any): Promise<any> {
+    const user = await this.authService.getUserById(req.user.id);
+    if (user.role === 'admin') {
+      const result = await this.reviewRepository.delete({ id });
+
+      if (result.affected === 0) {
+        throw new NotFoundException(`Review with ${id} not found`);
+      }
+      return {
+        success: true,
+        message: 'Review deleted Successfully',
+      };
+    } else {
+      throw new UnauthorizedException(
+        'UnAuthorized',
+        'You are not Authorized to access this endpoint ',
+      );
+    }
   }
 }
